+++ NEWS.md
@@ -1,4 +1,4 @@
-# santoku (development version)
+# santoku 0.9.0
 
 ## Breaking changes